remove commander (redundant)
