+++ package.json
@@ -23,7 +23,6 @@
     "arweave": "^1.10.3",
     "body-parser": "^1.19.0",
     "colors": "^1.4.0",
-    "commander": "^7.0.0",
     "dotenv": "^8.2.0",
     "event-stream": "^4.0.1",
     "express": "^4.17.1",
