+++ src/database/sync.database.ts
@@ -1,38 +1,37 @@
-import progress from 'progress';
-import { log } from '../utility/log.utility';
-import { block, currentBlock } from '../query/block.query';
-import { storeBlock } from './batch.database';
+import ProgressBar from 'progress';
+import {log} from '../utility/log.utility';
+import {block, currentBlock} from '../query/block.query';
+import {storeBlock} from './batch.database';
 
-export let bar: progress;
+export let bar: ProgressBar;
 export let startHeight = 0;
 
 export async function startSync() {
-    log.info(`[database] starting sync`);
-    
-    const initialBlock = await currentBlock();
-    startHeight = initialBlock.height;
-    bar = new progress(
-        `[database] :current/:total blocks synced [:bar] :percent :etas`,
-        {
-            complete: '|',
-            incomplete: ' ',
-            total: startHeight,
-        },
-    );
+  log.info(`[database] starting sync`);
 
-    await storeBlock(initialBlock);
-    traverseBlocks(startHeight - 1);
+  const initialBlock = await currentBlock();
+  startHeight = initialBlock.height;
+  bar = new ProgressBar(
+      `:current/:total blocks synced [:bar] :percent :etas`,
+      {
+        complete: '|',
+        incomplete: ' ',
+        total: startHeight,
+      },
+  );
+
+  await storeBlock(initialBlock);
+  traverseBlocks(startHeight - 1);
 }
 
 export async function traverseBlocks(height: number) {
-    const currentBlock = await block(height);
-    await storeBlock(currentBlock);
-    bar.tick(1);
+  const currentBlock = await block(height);
+  await storeBlock(currentBlock);
+  bar.tick(1);
 
-    if (height > 1) {
-        traverseBlocks(height - 1);
-    } else {
-        log.info(`[database] sync complete`);
-    }
-    
+  if (height > 1) {
+    traverseBlocks(height - 1);
+  } else {
+    log.info(`[database] sync complete`);
+  }
 }
